custom 404

+++ src/src/Controller/UserController.php
@@ -68,6 +68,10 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
+        if (!$user) {
+            return $this->redirectToRoute('app_error');
+        }
+
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -76,6 +80,10 @@ class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, UserRepository $userRepository): Response
     {
+        if (!$user) {
+            return $this->redirectToRoute('app_error');
+        }
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -94,6 +102,10 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, UserRepository $userRepository): Response
     {
+        if (!$user) {
+            return $this->redirectToRoute('app_error');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user, true);
         }
